New endpoints for CRUD operations on franchises

+++ src/main/kotlin/mabersold/plugins/Routing.kt
@@ -9,8 +9,8 @@ import mabersold.models.api.MetricType
 import mabersold.models.api.MetroData
 import mabersold.models.api.MetroOptions
 import mabersold.models.api.MetroReport
-import mabersold.models.api.requests.CreateLeagueRequest
-import mabersold.models.api.requests.UpdateLeagueRequest
+import mabersold.models.api.requests.SaveFranchiseRequest
+import mabersold.models.api.requests.SaveLeagueRequest
 import mabersold.services.FranchiseDataService
 import mabersold.services.LeagueDataService
 import mabersold.services.MetroDataService
@@ -73,36 +73,102 @@ fun Application.configureRouting() {
                 val leagues = leagueDataService.all()
                 call.respond(leagues)
             }
-            get("/{leagueId}/franchises") {
-                val leagueId = call.parameters["leagueId"]?.toInt() ?: throw IllegalArgumentException("Invalid League ID")
-                val franchises = franchiseDataService.getFranchises(leagueId)
-
-                call.respond(franchises)
-            }
             get("/{leagueId}/seasons") {
-                val leagueId = call.parameters["leagueId"]?.toInt() ?: throw IllegalArgumentException("Invalid League ID")
+                val leagueId = call.parameters["leagueId"]?.toIntOrNull()
+                if (leagueId == null) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "League ID is required")
+                    return@get
+                }
+
                 val seasons = seasonDataService.getSeasons(leagueId).sortedBy { it.startYear }
 
                 call.respond(seasons)
             }
             post {
-                val request = call.receive<CreateLeagueRequest>()
-                val created = leagueDataService.create(request)
+                val request = call.receive<SaveLeagueRequest>()
+                if (!request.canCreate()) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "Invalid request: league name and sport are required")
+                }
+
+                val created = leagueDataService.create(request.name!!, request.sport!!)
                 created?.let {
                     call.respond(it)
                 }
             }
             patch("/{leagueId}") {
-                val leagueId = call.parameters["leagueId"]?.toInt() ?: throw IllegalArgumentException("Invalid League ID")
-                val request = call.receive<UpdateLeagueRequest>()
+                val leagueId = call.parameters["leagueId"]?.toIntOrNull()
+                if (leagueId == null) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "League ID is required")
+                    return@patch
+                }
+
+                val request = call.receive<SaveLeagueRequest>()
                 val updated = leagueDataService.update(leagueId, request.name, request.sport)
                 updated?.let {
                     call.respond(it)
                 }
             }
             delete("/{leagueId}") {
-                val leagueId = call.parameters["leagueId"]?.toInt() ?: throw IllegalArgumentException("Invalid League ID")
+                val leagueId = call.parameters["leagueId"]?.toIntOrNull()
+                if (leagueId == null) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "League ID is required")
+                    return@delete
+                }
+
                 val isDeleted = leagueDataService.delete(leagueId)
+                if (isDeleted) {
+                    call.respond(HttpStatusCode.NoContent)
+                } else {
+                    call.respond(HttpStatusCode.BadRequest)
+                }
+            }
+        }
+        route("/franchises") {
+            get {
+                val leagueId = call.request.queryParameters["leagueId"]?.toIntOrNull()
+                if (leagueId == null) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "League ID is required")
+                    return@get
+                }
+
+                val franchises = franchiseDataService.getFranchises(leagueId)
+
+                call.respond(franchises)
+            }
+            post {
+                val request = call.receive<SaveFranchiseRequest>()
+                if (!request.canCreate()) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "Invalid request: name, is_defunct, and league_id must be provided")
+                    return@post
+                }
+
+                val created = franchiseDataService.createFranchise(request.name!!, request.isDefunct!!, request.leagueId!!)
+                created?.let {
+                    call.respond(it)
+                }
+            }
+            patch("/{franchiseId}") {
+                val franchiseId = call.parameters["franchiseId"]?.toIntOrNull()
+                if (franchiseId == null) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "Invalid or missing franchiseId")
+                    return@patch
+                }
+
+                val request = call.receive<SaveFranchiseRequest>()
+
+                franchiseDataService.updateFranchise(franchiseId, request.name, request.isDefunct, request.leagueId)?.let { updated ->
+                    call.respond(updated)
+                }
+            }
+            delete("/{franchiseId}") {
+                val franchiseId = call.parameters["franchiseId"]?.toIntOrNull()
+                if (franchiseId == null) {
+                    call.respond(HttpStatusCode.UnprocessableEntity, "Invalid or missing franchiseId")
+                    return@delete
+                }
+
+                val isDeleted = franchiseDataService.deleteFranchise(franchiseId)
+
                 if (isDeleted) {
                     call.respond(HttpStatusCode.NoContent)
                 } else {
